Markdown converter/requirements.txt
markitdown[all]
customtkinter==5.2.2
pillow==10.2.0
markdown2==2.4.12 